Rename huggingface-cli jobs to hf jobs
