+++ src/huggingface_hub/commands/huggingface_cli.py
@@ -18,7 +18,6 @@ from huggingface_hub.commands._cli_utils import show_deprecation_warning
 from huggingface_hub.commands.delete_cache import DeleteCacheCommand
 from huggingface_hub.commands.download import DownloadCommand
 from huggingface_hub.commands.env import EnvironmentCommand
-from huggingface_hub.commands.jobs import JobsCommands
 from huggingface_hub.commands.lfs import LfsCommands
 from huggingface_hub.commands.repo import RepoCommands
 from huggingface_hub.commands.repo_files import RepoFilesCommand
@@ -46,7 +45,6 @@ def main():
     DeleteCacheCommand.register_subcommand(commands_parser)
     TagCommands.register_subcommand(commands_parser)
     VersionCommand.register_subcommand(commands_parser)
-    JobsCommands.register_subcommand(commands_parser)
 
     # Experimental
     UploadLargeFolderCommand.register_subcommand(commands_parser)
